removing duplicate code and and parametrizing url
